requirements.txt
asttokens==2.2.1
attrs==23.1.0
azure-core==1.29.2
azure-storage-blob==12.17.0
backcall==0.2.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.2
Flask-Session==0.5.0
fonttools==4.42.0
greenlet==2.0.2
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
msal==1.23.0
nbformat==5.9.2
nest-asyncio==1.5.7
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.16.1
plotly-express==0.4.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyodbc==4.0.39
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.19
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.3
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
